% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_shifted.R
\name{geom_shifted}
\alias{geom_shifted}
\alias{dgeom_shifted_manual}
\alias{util_caller2}
\title{Geometric Distribution Supported on Positive Integers}
\usage{
dgeom_shifted_manual(x, prob, log = FALSE)

util_caller2(b = 5)
}
\arguments{
\item{prob}{probability of success in each trial. \code{0 < prob <= 1}.}
}
\description{
Geometric Distribution Supported on Positive Integers

#' @rdname geom_shifted
#'
#' @export
dgeom_shifted_programmatic <- shift_support(args = formals(dgeom),
distribution_function = dgeom,
shift_args = list("x = x -1"))
}
