% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/support_utils.R
\name{utils}
\alias{utils}
\alias{shift_support}
\alias{util2}
\alias{util3}
\alias{stats_util}
\title{Function-building utilities}
\usage{
shift_support(args, distribution_function, shift_args, env = parent.frame())

util2(a = 10)

util3(a = 10)

stats_util(distribution_function)
}
\arguments{
\item{args}{arguments passed by the user to the created function}

\item{distribution_function}{(...)}

\item{shift_args}{string describing transformations to inputs of the
distribution function, usually a 'rigid shift' indicating a different support}

\item{env}{environment in which to evaluate the new function}
}
\description{
Function-building utilities
}
